Refactor SchemaGenerator partials and tests
