+++ tests/SecureSpec.AspNetCore.Tests/SchemaGeneratorTestsReflection.cs
@@ -1,0 +1,12 @@
+using System.Reflection;
+using SecureSpec.AspNetCore.Schema;
+
+namespace SecureSpec.AspNetCore.Tests;
+
+internal static class SchemaGeneratorTestsReflection
+{
+    internal static MethodInfo GetApplyNullabilityMethod() =>
+        typeof(SchemaGenerator)
+            .GetMethod("ApplyNullability", BindingFlags.Instance | BindingFlags.NonPublic)
+        ?? throw new InvalidOperationException("ApplyNullability method not found.");
+}
